+++ HISTORY.rst
@@ -5,6 +5,10 @@ History
 1.2dev
 ------
 
+- Remove bootstrap macros for yafowil and set factory defaults for common
+  widgets where appropriate.
+  [rnix]
+
 - Update to bootstrap 3.2.
   [rnix]